Impl error decoding on concrete types instead of generic
